requirements.txt
allure-pytest==2.8.17
allure-python-commons==2.8.17
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
enum34==1.1.10
eulxml==1.1.3
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
lxml==4.5.2
more-itertools==8.4.0
namedlist==1.7
nose==1.3.7
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
ply==3.11
py==1.9.0
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==5.4.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
wcwidth==0.2.5
wer==0.1.0
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.1.0
